Update gotestsum to add timestamps to junit output
The current release of gotestsum is missing timestamps in the junit
data, which makes it difficult to import in an external system later.

https://github.com/gotestyourself/gotestsum/commit/012a85e34a7ce5554057d512e55dcb
includes the necessary changes to add the timestamp for the test run to
the junit output.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit bfbebf02b7c413c7afc9c253fb2199803c9e7c69)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>