[release/1.4] Update Go to 1.15.13
go1.15.13 (released 2021-06-03) includes security fixes to the archive/zip, math/big,
net, and net/http/httputil packages, as well as bug fixes to the linker, the go
command, and the math/big and net/http packages. See the Go 1.15.13 milestone on
the issue tracker for details:

https://github.com/golang/go/issues?q=milestone%3AGo1.15.13+label%3ACherryPickApproved

go1.15.12 (released 2021-05-06) includes a security fix to the net/http package,
as well as bug fixes to the runtime, the compiler, and the archive/zip, time,
and syscall packages. See the Go 1.15.12 milestone on the issue tracker for details:

https://github.com/golang/go/issues?q=milestone%3AGo1.15.12+label%3ACherryPickApproved

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>