[Vendor] Update hcsshim to 0.8.18
0.8.18 contains a fix for shim delete behavior, please see:
microsoft/hcsshim#1041

This time it brings in quite a few new files as 1.5 was on 0.8.16 of hcsshim
and 0.8.17 brought in a few new things. 0.8.18 doesn't actually bring in any new
vendored files but the version of the runhcs-shim built is based off of the commit
that's vendored into Containerd, so we need to vendor the version that has the fix.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit a81f05f3602a2b1eee9bff12ce9b985d16aadb60)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>