Add option to disallow pulling the same image in parallel for cri
Starting multiple pods on a single node, all using the same image and when the image isn't already present
on the node has pretty awful performance (especially on Windows, as unpack isn't cheap). This
change adds an option to skip this, by keeping an in memory cache of in progress pulls
at the cri layer.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>