log containerd's panic.log if one is present during startup
Currenetly (on windows) when there's an existing panic.log file
during containerd startup, it is renamed to panic.log.old and a
new panic.log file is created instead. This means, that in the
rare case of multiple containerd restarts, we can lose old panic
logs. To remedy that, log the contents of panic.log file before
renaming it.

Signed-off-by: Maksim An <maksiman@microsoft.com>