Fix Linux CI Linter using go 1.15.14
For some reason the Linux CI runs end up using go 1.15.14 instead of 1.16.6 for
the Windows runs, or any of the other CI steps. Not sure if this is due to
the linter installing it's own version of go or something else.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit 534685f9558cc67125156ba2fbdd0e0ade74cc33)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>