Test: revert hcsshim containerd dep to 1.4.x
There's an unfortunate series of events that occurs when trying to vendor
containerd 1.5.x into kubernetes/kubernetes because of the k8s.io deps in the
root go.mod. As hcsshim depends on containerd, we end up in the same rut. This is
just a test to see if anything fails CI wise when reverting hcsshims containerd dep to 1.4.
The only thing we use the dep for is the task definitions so shouldn't harm much as they haven't changed.
Tried the k8s e2e tests already but this is simpler to kick off a ci run to see if anything else explodes

Signed-off-by: Daniel Canter <dcanter@microsoft.com>