Allow docker pusher to follow redirects
http.Request#GetBody is called to get a new request body reader when an initial request returned a redirect.
The previous implementation only allowed the GetBody function to be called once, (which it used to set the body on the initial request)
causing redirects to fail when calling GetBody a second time.

This implementation caches the first read in a closure to allow subsequent GetBody calls to succeed during redirects.

Fixes containerd#5978

Signed-off-by: Kern Walster <walster@amazon.com>