Fix panic in metadata content writer on copy error
The `createAndCopy` function is only called when `nw.w` is nil
in order to create a new writer and prepare it. The current code
is attempting to close `nw.w` when there is a copy error. The
correct behavior would be to close the new writer and not touch `nw.w`.

Signed-off-by: Derek McGowan <derek@mcg.dev>
(cherry picked from commit b9cf0d75a9f4a8a09ab55caa1983a8988ba4834d)
Signed-off-by: Derek McGowan <derek@mcg.dev>