Add docs for ctr
Adding examples for some options for the ctr tool. This is not
exhaustive at the moment. This also does not show how to use the
checkpoint feature or snapshotting. This readme contains very basic
usage and not in a lot of details as this document would intend to
encourage users to view the help contents for subcommands.

Signed-off-by: prateeknischal <prateeknischal25@gmail.com>