cri: pass pull image credentials to snapshotter
Pass them via annotations so that a snapshotter can fetch image contents
with proper credentials. It is needed for remote snapshotters like
estargz and nydus. While it is possible for the snapshotters to parse
credentials on their own, but it several drawbacks:
1. snapshotter needs to access containerd config file
2. snapshotter needs to handle different credential sources that are
   already handled by kubelet/containerd
3. snapshotter needs additional kubernetes privileges to access
   kubernetes secrets APIs

By passing image pulling credentials to snapshotters, we can avoid the
above drawbacks. The behavior is still toggled by the current
`disable_snapshot_annotations` option and is disabled by default.

Signed-off-by: Peng Tao <bergwolf@hyper.sh>