metadata: make sure snapshots' children are removable
Sometimes I see snapshots that have children in "children" bucket,
but not in "snapshots" bucket. In this case, there is no way to remove
the snapshots.

This change adds an extra check to make sure that snapshots' children
are actually removable. If the child is deleted even partially,
it should not cause ErrFailedPrecondition.

Signed-off-by: Kazuyoshi Kato <katokazu@amazon.com>